result_analysis_project/requirements.txt
Flask
pandas
python-docx
